pyproject.toml
[tool.poetry]
authors = ["Marcus Gabriel <marcusgabriel.ds@gmail.com>"]
description = "A backend to manage different equipment of an FPSO (Floating Production, Storage and Offloading)"
name = "fpso-manager-backend"
version = "0.1.0"

[tool.poetry.dependencies]
django-ninja = "^0.12.3"
python = "^3.9"
python-decouple = "^3.4"

[tool.poetry.dev-dependencies]
prospector = "^1.3.1"
black = "^21.5b1"
django-test-without-migrations = "^0.6"
ipython = "^7.23.1"
model-bakery = "^1.3.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.isort]
atomic = true
combine_star = true
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
known_django = "django"
line_length = 100
lines_after_imports = 2
multi_line_output = 3
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip = ["venv", "env", "migrations", ".env", ".venv"]
use_parentheses = true

[tool.black]
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.venv
  | _build
  | build
  | dist
  | [a-z_]+/migrations
)/
'''
line-length = 100
quiet = true
target-version = ['py36']
